cs-onboarding/requirements.txt
Authlib==1.6.5
blinker==1.9.0
Flask-Limiter==3.9.2
boto3==1.40.57
botocore==1.40.57
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.0
Flask==3.1.2
Flask-WTF==1.2.2
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
httplib2==0.31.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.3
oauthlib==3.3.1
packaging==25.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.14.0
six==1.17.0
urllib3==2.5.0
Werkzeug==3.1.3
gunicorn==23.0.0
alembic==1.13.1
Flask-Caching==2.1.0
Flask-Compress==1.15
Flask-Talisman==1.1.0
python-magic==0.4.27
redis==5.0.1
sentry-sdk[flask]==1.40.0